feat(dashboard): full-height sidebar shell (Vercel-style layout)
Flips the app layout grid so the sidebar runs the full vertical
extent of the page and the TopHeader sits only over the content
column.

- AppLayout / OrgLayout: outer flex changes from column to row
- SidebarProvider: --header-offset defaults to 0 (was 3.5rem when not
  impersonating); only the impersonation banner shifts the sidebar
- TopHeader: stripped of logo, org slug, and user avatar — kept project
  switcher, breadcrumb anchor, Docs/Changelog links, plus a new
  SidebarTrigger
- New SidebarBrand mounts at the top of AppSidebar / OrgSidebar
- New SidebarUserMenu mounts at the bottom of both sidebars
- ThemeSwitcher moved into the user dropdown

Targets the org-home redesign branch — meant to land on top of that
work before merging the whole thing into main.
